requirements.txt
# torch
torchvision
tqdm
matplotlib
numpy
seaborn
torchaudio
librosa
# pip3 install torch torchvision
# pip3 install torchaudio
